Update DB Types

+++ src/db/global/activity_history.rs
@@ -8,7 +8,7 @@ pub struct Model {
     #[sea_orm(primary_key, auto_increment = true)]
     pub user_id: String,
     #[sea_orm(primary_key, auto_increment = false)]
-    pub activity: i64,
+    pub activity: i32,
     #[sea_orm(column_type = "Double")]
     pub time_played: f64,
 }
