+++ src/db/global/activity_user_data.rs
@@ -7,10 +7,10 @@ use serde::{Deserialize, Serialize};
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
     pub user_id: String,
-    pub last_played_activity: Option<i64>,
+    pub last_played_activity: Option<i32>,
     #[sea_orm(column_type = "Double", nullable)]
     pub last_played_time_played: Option<f64>,
-    pub current_activity: Option<i64>,
+    pub current_activity: Option<i32>,
     pub current_activity_start_time: Option<DateTime>,
 }
